+++ CHANGES.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### 0.1.4 (2014-01-22)
+
+* Added `StateCounter`, a state counter based on Redis `Hash`. To see an example of its usage, see [`tidehunter`](https://github.com/woozyking/tidehunter#example-2-without-limit).
+
 ### 0.1.3 (2014-01-20)
 
 * Added `QueueHandler`, inherits standard `logging.Handler` that `emit` to any standard `Queue` compatible implementations, including all the `Queue` implementations in this library